LLM/requirements.txt
anthropic>=0.30
datasets>=2.18
huggingface_hub>=0.23
llama-cpp-python>=0.2.77
peft>=0.11
pyyaml>=6.0
sentencepiece>=0.2
torch>=2.3
transformers>=4.44
trl>=0.9
unsloth>=2024.8
